+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Versions follow [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [0.8.12] - 2026-09-21
+
+### Fixed
+
+- A link in a Markdown document opens that file in a new tab. The link can
+  point at any project file, including one in a parent folder.
+
 ## [0.8.11] - 2026-09-17
 
 ### Added
@@ -446,7 +453,8 @@ Versions follow [Semantic Versioning](https://semver.org/).
   minimized regression test.
 - ADR-001…ADR-007 and the mdBook user-guide skeleton.
 
-[Unreleased]: https://github.com/mpakus/1537paperstreet/compare/v0.8.11...HEAD
+[Unreleased]: https://github.com/mpakus/1537paperstreet/compare/v0.8.12...HEAD
+[0.8.12]: https://github.com/mpakus/1537paperstreet/compare/v0.8.11...v0.8.12
 [0.8.11]: https://github.com/mpakus/1537paperstreet/compare/v0.8.10...v0.8.11
 [0.8.10]: https://github.com/mpakus/1537paperstreet/compare/v0.8.9...v0.8.10
 [0.8.9]: https://github.com/mpakus/1537paperstreet/compare/v0.8.8...v0.8.9